requirements.txt
aws-cdk-lib==2.24.0
constructs>=10.0.0,<11.0.0
flask-lambda
boto3